Struct stellar_sql::component::table::Table

pub struct Table {
    pub name: String,
    pub fields: HashMap<String, Field>,
    pub primary_key: Vec<String>,
    pub foreign_key: Vec<String>,
    pub reference_table: Option<String>,
    pub reference_attr: Option<String>,
    pub rows: Vec<Row>,
    pub is_data_loaded: bool,
    pub is_dirty: bool,
    pub dirty_cursor: u32,
    pub is_delete: bool,
    is_predicate_init: bool,
    row_set: HashSet<usize>,
    pub public_key: i32,
    uuid: String,
}

Fields
name: String fields: HashMap<String, Field> primary_key: Vec<String> foreign_key: Vec<String> reference_table: Option<String> reference_attr: Option<String> rows: Vec<Row> is_data_loaded: bool is_dirty: bool dirty_cursor: u32 is_delete: bool is_predicate_init: bool row_set: HashSet<usize> public_key: i32 uuid: String
Methods

impl Table

pub fn new(name: &str) -> Table

pub fn load_meta(
username: &str,
db_name: &str,
table_name: &str
) -> Result<Table, TableError>

Load a table with meta data

pub fn format_meta(&mut self, meta: TableMeta)

format metadata into table

pub fn load_rows_data(
&mut self,
username: &str,
db_name: &str
) -> Result<(), TableError>

load the particular range of rows from storage

pub fn load_all_rows_data(
&mut self,
username: &str,
db_name: &str
) -> Result<(), TableError>

load the all data from storage

pub fn insert_new_field(&mut self, field: Field)

pub fn insert_row(&mut self, row: Vec<(&str, &str)>) -> Result<(), TableError>

insert row into the table key and value are &str, and will be formated to the right type.

pub fn get_all_rows_set(&self) -> HashSet<usize>

return the set of all rows' id of the table

pub fn operator_filter_rows(
&mut self,
field_name: &str,
operator: &str,
value: &str
) -> Result<HashSet<usize>, TableError>

filter rows by the predicate and update the row_set

Note: this assume all data of rows and the predicate follow the rules, so there is no check for data type and field name.

pub fn set_row_set(&mut self, set: HashSet<usize>)

set the new row set

pub fn select(
&mut self,
field_names: Vec<String>
) -> Result<SelectData, TableError>

select fields from rows in row_set of the table