Module stellar_sql::sql::parser

Structs

Parser

Enums

ParserError

Functions

assert_token
Assert the symbol is the expected token.

check_id
Check if the symbol is an identifier

check_token
Check if the next symbol is the expected token.

get_id_list
Get a list of identifiers, which in form as

operator_priority

parse_infix_postfix
parse predicate tokens from infix to postfix

parse_postfix_tree
Parse a postfix to a binary tree, and do semantic check

parse_predicate
Parse a predicate as a tree

parse_select
Parse select query

parser_create_table

parser_insert_into_table